Added handle missing config errors and Personal Access Token docs
