+++ tests/Fixtures/Models/Company.php
@@ -15,13 +15,4 @@ class Company extends Authenticatable
     {
         return 'id';
     }
-
-    public static function createCompany()
-    {
-        \DB::table('companies')->insert([
-            'name' => 'Samuel',
-            'email' => 'sam.martins.dev@gmail.com',
-            'password' => \Hash::make('456'),
-        ]);
-    }
 }
