+++ tests/Unit/AuthConfigHelperTest.php
@@ -3,6 +3,7 @@
 namespace SMartins\PassportMultiauth\Tests\Unit;
 
 use SMartins\PassportMultiauth\Config\AuthConfigHelper;
+use SMartins\PassportMultiauth\Exceptions\MissingConfigException;
 use SMartins\PassportMultiauth\Tests\Fixtures\Models\Company;
 use SMartins\PassportMultiauth\Tests\Fixtures\Models\Customer;
 use SMartins\PassportMultiauth\Tests\Fixtures\Models\User;
@@ -19,22 +20,23 @@ class AuthConfigHelperTest extends TestCase
 
     public function testGetProviderGuard()
     {
-        $guard = AuthConfigHelper::getProviderGuard('companies');
+        $guard = AuthConfigHelper::getProviderGuard('companies', new Company);
 
         $this->assertEquals('company', $guard);
     }
 
     public function testGetProviderGuardWithNotPassportDriver()
     {
+        $this->expectException(MissingConfigException::class);
+        $this->expectExceptionMessage('Any guard with driver "passport" found to '.Customer::class. '. Please, check your config/auth.php file.');
+
         config(['auth.guards.customer.driver' => 'token']);
         config(['auth.guards.customer.provider' => 'customers']);
 
         config(['auth.providers.customers.driver' => 'eloquent']);
         config(['auth.providers.customers.model' => Customer::class]);
 
-        $guard = AuthConfigHelper::getProviderGuard('customers');
-
-        $this->assertNull($guard);
+        AuthConfigHelper::getProviderGuard('customers', new Customer);
     }
 
     public function testGetUserGuard()
@@ -53,27 +55,9 @@ class AuthConfigHelperTest extends TestCase
 
     public function testGetUserProviderWithModelNotExistentOnProviders()
     {
-        $provider = AuthConfigHelper::getUserProvider(new Customer);
+        $this->expectException(MissingConfigException::class);
+        $this->expectExceptionMessage('Any provider found to '.Customer::class.'. Please, check your config/auth.php file.');
 
-        $this->assertNull($provider);
-    }
-
-    /**
-     * Setup auth configs.
-     *
-     * @return void
-     */
-    protected function setAuthConfigs()
-    {
-        // Set up default entity
-        config(['auth.guards.api.driver' => 'passport']);
-        config(['auth.guards.api.provider' => 'users']);
-        config(['auth.providers.users.model' => User::class]);
-
-        // Set up company entity
-        config(['auth.guards.company.driver' => 'passport']);
-        config(['auth.guards.company.provider' => 'companies']);
-        config(['auth.providers.companies.driver' => 'eloquent']);
-        config(['auth.providers.companies.model' => Company::class]);
+        AuthConfigHelper::getUserProvider(new Customer);
     }
 }
